Add PHPStan static analysis, fix real bugs it and PHPStan surfaced
Sets up PHPStan (level 5, phpstan.neon) with a bootstrap stub for
DIR_ROOT/IS_DEV (real globals a consuming app's entry script always
defines, invisible to PHPStan analyzing this package in isolation) and
a baseline covering the 49 pre-existing issues not fixed here - mostly
Model/File.php's GD image-handling code (never fully adapted to PHP
8's resource->GdImage migration) and Model/Push/* (some of which
reference Appacman\* classes that don't exist in this package at all).
Deliberately not attempting those in this pass: security/correctness
sensitive, and not easily verifiable without real image/push fixtures.

Fixed along the way, each verified individually:
- Routing\Projects::$currentProject was a typed nullable property with
  no default - reading it via `== null` (the "nothing matched, fall
  back to default" check) threw "must not be accessed before
  initialization" instead of comparing falsy, whenever no configured
  domain matched the request at all. Same bug class as the
  Utils\Language fix from earlier in this modernization.
- Routing\Projects::getRegularExpression()'s `$pos == 0` check treated
  strpos()'s "not found" (false) the same as "found at position 0"
  (false == 0 is true in PHP) - silently populated the default-project
  fallback from whichever non-{lang} project was processed first,
  instead of only the one actually flagged isDefault. Together with
  the bug above, an unmatched request against a real config could
  silently resolve to the wrong sub-project instead of either matching
  correctly or throwing. Added tests/Routing/ProjectsTest.php - this
  class had zero coverage despite running on every request.
- Model\Encryptor\TwoWay::iv_bytes() was declared to return string but
  actually returned openssl_cipher_iv_length()'s int - worked today
  only because nothing in this codebase declares strict_types.
- Model\MySQL\PDO::restart() returned an int (from `$ok *= ...`
  boolean multiplication) instead of the declared bool.
- Model\Utils\StringUtils::validateURL()'s dead `?? ''` and
  Routing\RouteCompiler's redundant `=== '?'` check, both now-provably
  unreachable after earlier fixes in this modernization.
- Controller\Cache\Cache::$isCaching and PDO.php's two malformed
  `@return` PHPDoc comments (parse errors, not real bugs).

+++ src/Model/Utils/DateUtils.php
@@ -237,7 +237,11 @@ class DateUtils
     {
         $start = strtotime($endDate->format(self::FORMAT_TIMESTAMP_DB));
         $end   = strtotime($startDate->format(self::FORMAT_TIMESTAMP_DB));
-        return round(abs($end - $start) / 60, 2);
+        // round(..., 2) preserved for its exact prior behavior even though
+        // the 2 decimals never survived the (previously implicit) cast to
+        // int - not changing the return type since that would change what
+        // every caller actually receives
+        return (int) round(abs($end - $start) / 60, 2);
     }
 
     public static function timeAgo(string $startString): string
